refactor assistant chat surface
